requirements.txt
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
httplib2==0.14.0
jwcrypto==1.0
oauthlib==3.1.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymodm==0.4.3
pymongo==3.12.3
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-json-logger==2.0.2
simplejson==3.16.0
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8